Error message styling update and css refactoring
Error messaging ia failing accessibility because of the lighter red
coloured font on the white background didn't have enough contrast.

This change also includes updating the transaction completion header
with the turquoise background because that also fails accessibility.
However, the styles for this will shortly be merged into HMRC AF.
When they are I will create PR's removing the lcoal styles that are
included here.

In this commit:
• updated all instances of classes required for the new styling changes
• removed unnecessary styles from paye.css
• updated markpup for transaction completion header
• added local styles for transaction completion header
(to be removed once merged into HMRC AF)
• added local styles for error/form validation
• updated pages to use local error summary helper
• created local helper for input, to include new classes
• Added <div class="form-group"> to wrap from in directorDetails.scala.html
page to fix broken layout issue