Bump HMRC AF to latest version and update transaction completion
HMRC AF has a new release - 2.249.0 which includes updates to
colour variables, font size increase for transaction completion
header in it's current and new styles for the newer transaction
completion markup.

This commit bumps to the latest version of HMRC AF to adopt the
colour changes and updates the markup to use the new transaction
completion styles. The unnecessary local styles and assets in
place for transaction completion header have been removed to tidy
and refactor local assets.