Part of the DAC report for PAYE noticed we had some missing and unhelpful legends in our fieldsets, which they failed us on. This PR addrsses that and as part of that work I have refactored messages.
Refactor messages

Currently we have .title, .heading and .legend in messages to display
the title, heading and legend for all pages. While this makes logical
sense, these 3 text snippets are all the same and I don't think it
makes sense that there's 3 messages in conf/messages, when 1 can be
used to represent all 3. I have refactored and updated these and on
some pages updated broken links to the messages files, fixing other
issues as part of the work to check and update legends that are
failing DAC testing.